requirements.txt
sstreamlit
python-dotenv
openai
langchain
langchain-openai
langchain-community
PyMuPDF
